feat(solution): adds part 1 & 2 day 13
Struggled a lot for nothing, I didn't realized that my calculation
of the height of the paper was wrong, I was assuming we will have
some dots on the last line, but as we were not having any, while
folding with x,497 we were not comparing the right lines 🤦

Also I need to have the reflex to build the algo using sparse
structure. Here after seeing some other solutions, it would have
make way more sense to use sparse sets to describe lines, or either
to just keep coordinates and to transpose them.