feat(solution): part1 day 15
When we have a better cost we re-analyze the coordinate. This is not
ideal, of course this is working, but it means that some paths will
be explored multiple times.
Need to think of a better way to not redo any path, I will have a
better working brain tomorrow morning.